Time: 1 ms (100%), Space: 45.9 MB (44.17%) - LeetHub

+++ 0260-single-number-iii/0260-single-number-iii.java
@@ -3,7 +3,7 @@ class Solution {
         int xor = 0;
         for(int i = 0; i < nums.length; i ++)
             xor = (xor ^ nums[i]);
-        long rightmost = ((xor & (xor - 1)) ^ xor); //must be long, bcoz if xor is the smallest integer, then xor - 1 will overflow the capacity of integer! 
+        int rightmost = ((xor & (xor - 1)) ^ xor); //must be long, bcoz if xor is the smallest integer, then xor - 1 will overflow the capacity of integer! 
         int bucket_1 = 0, bucket_2 = 0;
         for(int i = 0; i < nums.length; i ++) {
             if((rightmost & nums[i]) != 0) //some non-zero number
